requirements.txt
aiogram>=3.4.0
aiohttp>=3.9.0
SQLAlchemy>=2.0.0
apscheduler>=3.10.0
Pillow>=10.0.0
matplotlib>=3.8.0
mistralai>=1.0.0
python-dotenv>=1.0.0
